Baseline for AWS::SES::ReceiptFilter with latest plugin changes
