+++ aws-ses-receiptfilter/aws-ses-receiptfilter.json
@@ -57,45 +57,33 @@
         "DueDate"
     ],
     "readOnlyProperties": [
-        "/properties/tps_code"
+        "/properties/TPSCode"
     ],
-    "identifiers": [
-        [
-            "/properties/tps_code"
-        ]
+    "primaryIdentifier": [
+        "/properties/TPSCode"
     ],
     "handlers": {
         "create": {
-            "protocolVersion": "1.0",
-            "runtimeType": "Lambda",
             "permissions": [
                 "initech:CreateReport"
             ]
         },
         "read": {
-            "protocolVersion": "1.0",
-            "runtimeType": "Lambda",
             "permissions": [
                 "initech:DescribeReport"
             ]
         },
         "update": {
-            "protocolVersion": "1.0",
-            "runtimeType": "Lambda",
             "permissions": [
                 "initech:UpdateReport"
             ]
         },
         "delete": {
-            "protocolVersion": "1.0",
-            "runtimeType": "Lambda",
             "permissions": [
                 "initech:DeleteReport"
             ]
         },
         "list": {
-            "protocolVersion": "1.0",
-            "runtimeType": "Lambda",
             "permissions": [
                 "initech:ListReports"
             ]
