+++ aws-ses-receiptfilter/pom.xml
@@ -5,7 +5,7 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.aws.ses.receiptfilter</groupId>
+    <groupId>com.amazonaws.ses.receiptfilter</groupId>
     <artifactId>aws-ses-receiptfilter-handler</artifactId>
     <name>aws-ses-receiptfilter-handler</name>
     <version>1.0-SNAPSHOT</version>
@@ -155,6 +155,7 @@
                         <exclude>**/BaseConfiguration*</exclude>
                         <exclude>**/BaseHandler*</exclude>
                         <exclude>**/HandlerWrapper*</exclude>
+                        <exclude>**/ResourceModel*</exclude>
                     </excludes>
                 </configuration>
                 <executions>
@@ -182,11 +183,6 @@
                                     <limits>
                                         <limit>
                                             <counter>BRANCH</counter>
-                                            <value>COVEREDRATIO</value>
-                                            <minimum>0.8</minimum>
-                                        </limit>
-                                        <limit>
-                                            <counter>LINE</counter>
                                             <value>COVEREDRATIO</value>
                                             <minimum>0.8</minimum>
                                         </limit>
