+++ aws-ses-receiptfilter/src/main/java/com/amazonaws/ses/receiptfilter/Configuration.java
@@ -1,0 +1,17 @@
+package com.amazonaws.ses.receiptfilter;
+
+import java.io.InputStream;
+import org.json.JSONObject;
+import org.json.JSONTokener;
+
+class Configuration extends BaseConfiguration {
+
+    public Configuration() {
+        super("aws-ses-receiptfilter.json");
+    }
+
+    public JSONObject resourceSchemaJSONObject() {
+        return new JSONObject(new JSONTokener(this.getClass().getClassLoader().getResourceAsStream(schemaFilename)));
+    }
+
+}
