--- aws-ses-receiptfilter/src/main/java/com/aws/ses/receiptfilter/Configuration.java
@@ -1,15 +1,0 @@
-package com.aws.ses.receiptfilter;
-
-import java.io.InputStream;
-
-class Configuration extends BaseConfiguration {
-
-    public Configuration() {
-        super("aws-ses-receiptfilter.json");
-    }
-
-    public InputStream resourceSchema() {
-        return this.getClass().getClassLoader().getResourceAsStream(schemaFilename);
-    }
-
-}
