+++ aws-ses-receiptfilter/template.yml
@@ -1,0 +1,22 @@
+AWSTemplateFormatVersion: "2010-09-09"
+Transform: AWS::Serverless-2016-10-31
+Description: AWS SAM template for the AWS::SES::ReceiptFilter resource type
+
+Globals:
+  Function:
+    Timeout: 60  # docker start-up times can be long for SAM CLI
+
+Resources:
+  TypeFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      Handler: com.amazonaws.ses.receiptfilter.HandlerWrapper::handleRequest
+      Runtime: java8
+      CodeUri: ./target/aws-ses-receiptfilter-handler-1.0-SNAPSHOT.jar
+
+  TestEntrypoint:
+    Type: AWS::Serverless::Function
+    Properties:
+      Handler: com.amazonaws.ses.receiptfilter.HandlerWrapper::testEntrypoint
+      Runtime: java8
+      CodeUri: ./target/aws-ses-receiptfilter-handler-1.0-SNAPSHOT.jar
